// projects/coach-hub/src/app/modules/coach/modules/programs/models/create-program.interface.ts
import { CreateProgramPriceInterface } from '../../../../app/models/create-program-price.interface';

export interface CreateProgramInterface {
  program_title: string;
  program_description: string;
  category: number;
  registration_start: string;
  registration_end: string;
  program_start: string;
  program_end: string;
  location_id: number;
  tags: number[];
  prices: CreateProgramPriceInterface[];
}